Damage system for player and enemy
GameManager.cs:
Added PlayerMissesEnemy() method, which reduces player's life by 1 if player missed 5 enemies.
Added the enemyDestroyed variable and a check in the EnemyHitsPlayer() method to allow the enemy to decrease the player's life only if the enemy is not destroyed.
The enemyIndex and fireIndex variables are removed, they are not needed. Instead of enemyIndex the EnemyHitsPlayer() or PlayerMissesEnemy() methods are called, where the index is passed as a parameter.

EnemyController.cs
Added checks in IEnumerator SlowEnemyMovement() if the enemy is destroyed. If the enemy is destroyed, the coroutine is stopped. Similarly with fire in coroutine SlowFire() in PlayerController.cs.
 In IEnumerator CallEnemy() the variable enemyDestroyed changes to false with every new enemy cycle.

Added the DestroyEnemy.cs script, which changes the value of the enemyDestroyed variable to true and disables the enemy when fire and enemy collide.