Create a new damage system for the player
The previous system was based on player and enemy indices, which caused the player to take damage if he killed a player and stayed put.
The current logic is based on collision. A fourth line of enemies is created with the picture turned off so that their collider hits the player's collider. Also created an animation that plays when the player collides with an enemy.