Fix the logic when a player misses enemies
Created a script that tracks whether an enemy from the fourth line collides with a player or not. If it doesn't, the enemy is considered missing. Previously, the enemy from the fourth line could appear even after the player destroyed the enemy, and then the enemy was mistakenly considered missed.