Add post-processing to scenes
I implemented a post-processing profile to add chromatic aberration and grain effects to the Camera objects. Additionally, I added the PostProcessing layer to the Camera objects and set the Render mode in Cameras' Canvas to Screen Space - Camera to ensure that the effects are applied to the UI as well. Finally, I created a post-process layer and volume for each camera in every scene, ensuring consistent effects throughout the game.

These changes enhance the overall visual quality and create a cohesive look across scenes, contributing to a more polished gameplay experience.

Files modified:
- TagManager.asset
- MainScene.unity
- StartMenuScene.unity